Test2 3

+++ Ev-System/src/main/java/com/example/Ev/System/repository/ServiceCenterRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ServiceCenterRepository extends JpaRepository<ServiceCenter, Integer> {
+public interface ServiceCenterRepository extends JpaRepository<ServiceCenter, Long> {
     Optional<ServiceCenter> findById(Integer id);
 }
